refactor: extract testimonials component from reviewIframe for better code organization
