+++ src/scss/_utility.scss
@@ -5,10 +5,6 @@
     border-color: var(--custom-border) !important;
 }
 
-.border-right-color {
-    border-right: 1px solid var(--custom-border) !important;
-}
-
 /** Hover effext for mcp or integrations apps */
 .hover-bg-grey-100-text-black {
     &:hover {
